# [CUDA] fix illegal memory access in attention