--- a/aten/src/ATen/native/transformers/cuda/attention_backward.cu
+++ b/aten/src/ATen/native/transformers/cuda/attention_backward.cu
@@ -648,8 +648,8 @@ _efficient_attention_backward(
     p.custom_mask_type = custom_mask_type;
     p.scale = sdp::calculate_scale(query, scale).expect_float();
     if (cu_seqlens_q.has_value()) {
-      p.cu_seqlens_q_ptr = (const int64_t*)cu_seqlens_q->const_data_ptr();
-      p.cu_seqlens_k_ptr = (const int64_t*)cu_seqlens_k->const_data_ptr();
+      p.cu_seqlens_q_ptr = (const int32_t*)cu_seqlens_q->const_data_ptr();
+      p.cu_seqlens_k_ptr = (const int32_t*)cu_seqlens_k->const_data_ptr();
     }
     if (window_size.has_value()) {
       p.window_size = *window_size;
@@ -750,7 +750,7 @@ _efficient_attention_backward(
       // when we need a staging area for gK/gV. let's avoid that
       if (Kernel::kNeedsAccumGradK || Kernel::kNeedsAccumGradV) {
         p.num_splits_key = std::min(
-            int(p.num_splits_key), 200 / ((int)(p.num_batches * p.num_heads)));
+            int32_t(p.num_splits_key), 200 / ((int32_t)(p.num_batches * p.num_heads)));
       }
     }
     if (!Kernel::kEnableSplitKeys || p.num_splits_key < 1) {