Fixes #150054

Didn't add a test since it takes a lot of memory and time(on cpu) to calculate this.