@Isalia20 2025-06-05
That should have fixed the tests. These are the benchmarks:

| seq_len | embed_dim |   old   |   new   |
|---------|-----------|---------|---------|
| 256     | 64        | 0.31 ms | 0.39 ms |
| 256     | 128       | 0.30 ms | 0.30 ms |
| 256     | 256       | 0.30 ms | 0.30 ms |
| 256     | 512       | 0.30 ms | 0.29 ms |
| 512     | 64        | 0.32 ms | 0.31 ms |
| 512     | 128       | 0.31 ms | 0.31 ms |
| 512     | 256       | 0.31 ms | 0.31 ms |
| 512     | 512       | 0.29 ms | 0.32 ms |
| 1024    | 64        | 0.40 ms | 0.43 ms |
| 1024    | 128       | 0.39 ms | 0.43 ms |
| 1024    | 256       | 0.39 ms | 0.43 ms |
| 1024    | 512       | 0.42 ms | 0.46 ms |
| 2048    | 64        | 0.73 ms | 0.78 ms |
| 2048    | 128       | 0.73 ms | 0.79 ms |
| 2048    | 256       | 0.80 ms | 0.78 ms |
| 2048    | 512       | 0.87 ms | 0.86 ms |
| 4096    | 64        | 2.47 ms | 2.60 ms |
| 4096    | 128       | 2.45 ms | 2.50 ms |
| 4096    | 256       | 2.48 ms | 2.51 ms |
| 4096    | 512       | 2.91 ms | 2.96 ms |